Increase performance of uncached module resolution
When loading a module, look in the cache to see if we have already
loaded one of its ancestors. If we have then we can resolve the
module path without having to call nvim_get_runtime_file() or
globpath(), both of which are very slow.

This significantly increases the performance of module resolution of
uncached loads. If there is already a cache file, then this has no
effect.